Write optimize includes logic in OptimizeIncludes() function. Assemble all arguments and scripts to pass to process argumets.
